+++ tailbench/harness/tbench_server.h
@@ -31,7 +31,7 @@ void tBenchServerFinish();
 
 size_t tBenchRecvReq(void** data);
 
-void tBenchSendResp(const void* data, size_t size, int32_t feature1, int32_t feature2, int32_t feature3);
+void tBenchSendResp(const void* data, size_t size, uint64_t feature1, uint64_t feature2, uint64_t feature3);
 
 #ifdef __cplusplus 
 }
